requirements.txt
appnope==0.1.0
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
Django==3.0.8
wheel==0.34.2
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
django-filter=2.3.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pip=20.0.2
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
PyYAML==5.3.1
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
setuptools=44.0.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
